+++ spring-security-authorization/src/main/java/ru/otus/domain/User.java
@@ -22,5 +22,5 @@ public class User {
 
     private String email;
 
-    private String role;
+    private String authority;
 }
