+++ spring-security-authorization/src/main/resources/application.yml
@@ -7,8 +7,8 @@ spring:
   sql:
     init:
       mode: always
-#      data-locations: data.sql
-#      schema-locations: schema.sql
+      data-locations: classpath:data.sql
+      schema-locations: classpath:schema.sql
   jpa:
     open-in-view: false
     show-sql: true
@@ -18,6 +18,6 @@ spring:
     hiddenmethod:
       filter:
         enabled: true
-#logging:
-#  level:
-#    ROOT: ERROR
+logging:
+  level:
+    ROOT: ERROR
