+++ spring-security-authorization/src/main/resources/schema.sql
@@ -32,5 +32,5 @@ create table Users (
 	username varchar not null primary key,
 	password varchar not null,
 	email varchar not null,
-	role varchar not null
+	authority varchar not null
 );
